requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# RTSP & Video Processing
opencv-python==4.9.0.80
av==11.0.0
numpy==1.24.3

# YOLOv11 & Detection
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2

# Multi-Object Tracking
filterpy==1.4.5
scipy==1.11.4
scikit-learn==1.3.2

# ANPR/OCR
easyocr==1.7.0
Pillow==10.2.0

# PTZ Control (ONVIF)
onvif-zeep==0.2.12
zeep==4.2.1

# Message Bus & Cache
redis==5.0.1
hiredis==2.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Object Storage
boto3==1.34.20
minio==7.2.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.20.0
